+++ db/data-generators/generate-json-data.js
@@ -1,11 +1,9 @@
 // this file creates 3 csvs that are stored in the local data folder
 // use the postgres or cassandra loader files to load the csv into a db
 const fs = require('fs');
+const colorList = require('./colors.js');
 
 const getRandomInt = (min, max) => Math.floor(Math.random() * ((max - min) + 1)) + min;
-
-// generate products
-console.log('starting products');
 
 const text = `mumblecore migas semiotics
   shabby-chic sustainable art-party blue-bottle ugh health goth activated-charcoal
@@ -22,67 +20,50 @@ const text = `mumblecore migas semiotics
 let titleWords = text.replace(/\n/g, '').replace('  ', '').split(' ');
 titleWords = titleWords.filter(word => word !== '');
 
-let csvContent = '';
+let allData = '';
+
 for (let i = 1; i <= 10000000; i += 1) {
-  const row = [];
+  const row = {};
+  row.id = i;
 
   // create a title
   const title = [];
   for (let j = 0; j < 5; j += 1) {
     title.push(titleWords[getRandomInt(0, titleWords.length - 1)]);
   }
-  row.push(`"${title.join(' ').trim()}"`);
+  row.title = title.join(' ').trim();
 
   // get main & hover images
-  row.push(`"https://s3-us-west-1.amazonaws.com/rapid-retail-product-images/${getRandomInt(0, 999)}.jpg"`);
-  row.push(`"https://s3-us-west-1.amazonaws.com/rapid-retail-product-images/${getRandomInt(0, 999)}.jpg"`);
+  row.main_img = `https://s3-us-west-1.amazonaws.com/rapid-retail-product-images/${getRandomInt(0, 999)}.jpg`;
+  row.hov_img = `https://s3-us-west-1.amazonaws.com/rapid-retail-product-images/${getRandomInt(0, 999)}.jpg`;
 
   // get price
-  row.push(getRandomInt(13, 29) * 5);
+  row.price = getRandomInt(13, 29) * 5;
 
-  csvContent += `${row.join(',')}\r\n`;
+  // get colors
+  const colors = [];
+  for (let j = 0; j < 4; j += 1) {
+    colors.push(colorList[getRandomInt(0, colorList.length - 1)]);
+  }
+
+  row.colors = colors;
+
+  // get related items
+  const related = [];
+  for (let j = 0; j < 4; j += 1) {
+    related.push(getRandomInt(1, 10000000));
+  }
+
+  row.related = related;
+
+  allData += `${JSON.stringify(row)}\r\n`;
 
   // append to file every 100000 rows
   if (i % 100000 === 0) {
-    fs.appendFileSync('db/data/products.csv', csvContent);
-    csvContent = '';
+    fs.appendFileSync('db/data/allData.json', allData);
+    allData = '';
     console.log(`completed ${i} products`);
   }
 }
 
-console.log('generated products');
-
-// generate products_colors
-console.log('starting products_colors');
-csvContent = '';
-
-for (let i = 1; i <= 10000000; i += 1) {
-  for (let j = 0; j < 4; j += 1) {
-    csvContent += `${i},${getRandomInt(1, 116)}\r\n`;
-  }
-
-  if (i % 100000 === 0) {
-    fs.appendFileSync('db/data/products_colors.csv', csvContent);
-    csvContent = '';
-    console.log(`completed ${i} products`);
-  }
-}
-
-console.log('generated products_colors!');
-
-// generate product_relations
-console.log('starting product_relations');
-csvContent = '';
-for (let i = 1; i <= 10000000; i += 1) {
-  for (let j = 0; j < 4; j += 1) {
-    csvContent += `${i},${getRandomInt(1, 10000000)}\r\n`;
-  }
-
-  if (i % 100000 === 0) {
-    fs.appendFileSync('db/data/product_relations.csv', csvContent);
-    csvContent = '';
-    console.log(`completed ${i} products`);
-  }
-}
-
-console.log('generated product_relations!');
+console.log('file generated');
